Fix compilation issues and reporting crash after STOP

+++ native_player/AlsaAudioEmitter.cpp
@@ -327,7 +327,8 @@ void AlsaAudioEmitter::workerThread(std::stop_token token) {
     setupAudioFormat(streamInfo.value().format);
     const auto dataToRequest =
         snd_pcm_frames_to_bytes(pcmHandle, bufferSize / 2);
-    auto dataAvailable = inputNode->waitForData(token, dataToRequest);
+    auto dataAvailable =
+        static_cast<size_t>(inputNode->waitForData(token, dataToRequest));
     if (dataAvailable < dataToRequest) {
       std::cerr << "Not enough data available for playback - requested "
                 << dataToRequest << " received " << dataAvailable
